orionis/src/requirements.txt
flask
functions-framework
google-cloud-datastore
pydantic
google-cloud-storage
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
google-generativeai==0.8.3
requests
python-dotenv
python-multipart
black
pyjwt
opencv-python
email-validator
cryptography
google-cloud-tasks
regex
instructor
xlsxwriter
flake8
mypy
pandas-stubs
types-requests
flask_swagger_ui
pytest
python-socketio
google-cloud-aiplatform
google-cloud-pubsub
sentry-sdk
google-play-scraper
itunespy
moviepy
requests
stripe